Update um-actions-form.php
Regarding field limit error prompts, this mod accounts for users who wish to use a label, a placeholder, both, or neither.  The label obviously takes precedence.  It shouldn't be much trouble to add an admin radio for choosing which to use in the error prompt when both are present, but this works for me, so I quit while I was ahead.  XD